requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
anyio==3.7.1
app-store-server-library==1.1.0
asn1==2.7.0
async-timeout==4.0.3
attrs==23.1.0
audioread==3.0.1
beautifulsoup4==4.12.2
blinker==1.8.2
CacheControl==0.13.1
cachetools==5.3.2
cattrs==23.1.2
certifi==2023.5.7
cffi==1.16.0
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
contourpy==1.1.1
cryptography==41.0.7
cycler==0.12.1
dataclasses-json==0.5.14
decorator==5.1.1
distlib==0.3.6
distro==1.8.0
dnspython==2.6.1
email_validator==2.1.1
emoji==2.8.0
enum-compat==0.0.3
exceptiongroup==1.1.3
fastapi==0.111.0
fastapi-cli==0.0.4
filelock==3.12.2
filetype==1.2.0
firebase-admin==6.2.0
Flask==3.0.3
fonttools==4.43.1
frozenlist==1.4.0
google-api-core==2.19.0
google-api-python-client==2.109.0
google-auth==2.29.0
google-auth-httplib2==0.1.1
google-cloud-core==2.3.3
google-cloud-firestore==2.13.1
google-cloud-storage==2.16.0
google-cloud-texttospeech==2.16.3
google-crc32c==1.5.0
google-resumable-media==2.6.0
googleapis-common-protos==1.61.0
grpcio==1.59.3
grpcio-status==1.59.3
h11==0.14.0
httpcore==0.18.0
httplib2==0.22.0
httptools==0.6.1
httpx==0.25.0
idna==3.4
importlib-resources==6.1.0
importlib_metadata==7.1.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.3.2
kiwisolver==1.4.5
langchain==0.0.283
langsmith==0.0.33
lazy_loader==0.3
librosa==0.10.1
llvmlite==0.41.1
lxml==4.9.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.1
matplotlib==3.8.0
mdurl==0.1.2
msgpack==1.0.7
multidict==6.0.4
mypy-extensions==1.0.0
nltk==3.8.1
numba==0.58.1
numexpr==2.8.5
numpy==1.25.2
openai==1.30.1
orjson==3.10.3
outcome==1.2.0
packaging==23.1
Pillow==10.1.0
pipenv==2023.7.11
platformdirs==3.9.1
pooch==1.7.0
proto-plus==1.22.3
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==2.3.0
pydantic_core==2.6.3
Pygments==2.18.0
PyJWT==2.8.0
pyOpenSSL==24.1.0
pyparsing==3.1.1
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-magic==0.4.27
python-multipart==0.0.9
PyYAML==6.0.1
regex==2023.8.8
requests==2.31.0
rich==13.7.1
rsa==4.9
scikit-learn==1.3.1
scipy==1.11.3
selenium==4.12.0
shellingham==1.5.4
sniffio==1.3.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.5
soxr==0.3.7
SQLAlchemy==2.0.20
starlette==0.37.2
tabulate==0.9.0
tenacity==8.2.3
threadpoolctl==3.2.0
tiktoken==0.5.1
tqdm==4.66.1
trio==0.22.2
trio-websocket==0.10.4
typer==0.12.3
typing-inspect==0.9.0
typing_extensions==4.8.0
ujson==5.10.0
unstructured==0.10.12
uritemplate==4.1.1
urllib3==2.0.4
uvicorn==0.24.0.post1
uvloop==0.19.0
virtualenv==20.24.0
virtualenv-clone==0.5.7
watchfiles==0.21.0
websockets==12.0
Werkzeug==3.0.3
wsproto==1.2.0
yarl==1.9.2
zipp==3.17.0
